refactor: move HardhatDB to hardhat-core
