+++ crates/rethnet_evm_napi/package.json
@@ -16,12 +16,8 @@
     "devDependencies": {
         "@napi-rs/cli": "^2.11.4",
         "chai": "^4.3.6",
+        "@nomicfoundation/ethereumjs-util": "^8.0.0",
         "mocha": "^10.0.0",
         "typescript": "~4.5.2"
-    },
-    "dependencies": {
-        "@nomicfoundation/ethereumjs-blockchain": "^6.0.0",
-        "@nomicfoundation/ethereumjs-statemanager": "^1.0.0",
-        "@nomicfoundation/ethereumjs-util": "^8.0.0"
     }
 }
